添加数据库相关类 & 修改 nat_type

+++ transmission/TransmissionProxy.cpp
@@ -75,7 +75,7 @@ TransmissionData TransmissionProxy::read(){
     if(strBytes.size() != 1)
         return ret;
 
-    int lengthOfSize = strBytes[0];
+    size_t lengthOfSize = strBytes[0];
     strBytes = m_socket->read(lengthOfSize);
     if(strBytes.size() != lengthOfSize)
         return ret;
